Refactor precode.go to main.go
